+++ libraries/base/changelog.md
@@ -41,6 +41,8 @@
 
   * Add `resizeSmallMutableArray#` to `GHC.Exts`.
 
+  * Add `Foldable1` type-class for non-empty structures that can be folded
+
 ## 4.13.0.0 *TBA*
   * Bundled with GHC *TBA*